fix: change type int to double for divide

+++ Program.cs
@@ -7,13 +7,13 @@
             Console.WriteLine("Hesap Makinesi");
 
             Console.Write("Birinci Sayıyı Giriniz: ");
-            int sayi1 = int.Parse(Console.ReadLine());
+            double sayi1 = double.Parse(Console.ReadLine());
 
             Console.Write("İşlem Seçiniz (+) (-) (*) (/) (%): ");
             string islem = Console.ReadLine();
 
             Console.Write("İkinci Sayıyı Giriniz: ");
-            int sayi2 = int.Parse(Console.ReadLine());
+            double sayi2 = double.Parse(Console.ReadLine());
 
             switch (islem)
             {
